Harden modular CLI install paths

+++ scripts/commands/install.sh
@@ -25,7 +25,7 @@ command_main() {
       sudo install -m 0755 "$LLM_SM_ENTRYPOINT" "$target_bin"
     fi
     if [[ "$source_root" != "$lib_target" ]]; then
-      sudo rm -rf -- "$lib_target/lib" "$lib_target/commands"
+      sudo rm -rf -- "${lib_target:?}/lib" "${lib_target:?}/commands"
       sudo cp -R -- "$source_root/lib" "$source_root/commands" "$lib_target/"
       sudo chmod -R a+rX "$lib_target"
     fi
@@ -34,7 +34,7 @@ command_main() {
       install -m 0755 "$LLM_SM_ENTRYPOINT" "$target_bin"
     fi
     if [[ "$source_root" != "$lib_target" ]]; then
-      rm -rf -- "$lib_target/lib" "$lib_target/commands"
+      rm -rf -- "${lib_target:?}/lib" "${lib_target:?}/commands"
       cp -R -- "$source_root/lib" "$source_root/commands" "$lib_target/"
       chmod -R a+rX "$lib_target"
     fi
